LayoutTests/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-resolved-min-size-auto-expected.txt


PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="block-box"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="inline-box"
FAIL Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="block-with-valid-aspect-ratio" assert_equals: block-with-valid-aspect-ratio: 'min-width:auto' should round-trip through getComputedStyle (as initial value) expected "auto" but got "0px"
FAIL Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="block-with-degenerate-aspect-ratio" assert_equals: block-with-degenerate-aspect-ratio: 'min-width:auto' should round-trip through getComputedStyle (as initial value) expected "auto" but got "0px"
FAIL Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="block-with-two-part-valid-aspect-ratio" assert_equals: block-with-two-part-valid-aspect-ratio: 'min-width:auto' should round-trip through getComputedStyle (as initial value) expected "auto" but got "0px"
FAIL Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="block-with-two-part-degenerate-aspect-ratio" assert_equals: block-with-two-part-degenerate-aspect-ratio: 'min-width:auto' should round-trip through getComputedStyle (as initial value) expected "auto" but got "0px"
FAIL Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="inline-with-valid-aspect-ratio" assert_equals: inline-with-valid-aspect-ratio: 'min-width:auto' should round-trip through getComputedStyle (as initial value) expected "auto" but got "0px"
FAIL Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="inline-with-degenerate-aspect-ratio" assert_equals: inline-with-degenerate-aspect-ratio: 'min-width:auto' should round-trip through getComputedStyle (as initial value) expected "auto" but got "0px"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="flex-item-row"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="flex-item-column"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="grid-item"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="display-none"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="display-none-valid-aspect-ratio"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="display-none-subtree-valid-aspect-ratio"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="display-none-subtree-flex-item"
PASS Resolved value of min-width & min-height 'auto' keyword behaves as expected on element with id="display-none-subtree-grid-item"
